requirements.txt
discord.py==1.7.3
dnspython==2.1.0
pymongo==3.12.1
python-dotenv==0.19.1